# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librarie API",
        "version": "0.1"
    },
    "paths": {
        "/carti-comanda": {
            "get": {
                "tags": [
                    "Carti-comanda"
                ],
                "summary": "Returneaza toate cartile comenzii",
                "description": "Returneaza toate cartile comenzii",
                "operationId": "5d0f4ede214fe8260fb9c5fa790fb1ee",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Carti-comanda"
                ],
                "summary": "Update cartea comenzii cu id-ul dat",
                "description": "Update cartea comenzii cu id-ul dat",
                "operationId": "47616788bf27de2fba42fffb17d9f7dc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_carte": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_comanda": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantitate": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "description": "",
                                        "type": "float",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Carti-comanda"
                ],
                "summary": "Adauga cartea comenzii",
                "description": "Adauga cartea comenzii",
                "operationId": "582c0116bb883f91549d831fbcecdc2a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_carte": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_comanda": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantitate": {
                                        "description": "",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subtotal": {
                                        "description": "",
                                        "type": "float",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/carti-comanda/{id}": {
            "get": {
                "tags": [
                    "Carti-comanda"
                ],
                "summary": "Returneaza cartea comenzii cu id-ul dat",
                "description": "Returneaza cartea comenzii cu id-ul dat",
                "operationId": "a9f31ae072ab093fc03eb9e73075d6b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii comenzii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Carti-comanda"
                ],
                "summary": "Sterge cartea comenzii cu id-ul dat",
                "description": "Sterge cartea comenzii cu id-ul dat",
                "operationId": "2e5d09713548516bab46b6a795944ff4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id-ul cartii comenzii dorite",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}